完成项目开发

+++ purchase-backend/purchase-admin/src/main/java/com/shanzhu/purchase/service/CkDepositoryOutService.java
@@ -41,6 +41,6 @@ public interface CkDepositoryOutService {
     /**
      * 通过编号查询 source_number
      */
-    List<CkmdDepositoryOut> getRowInfoByPurchaseNumber(int number);
+    List<CkmdDepositoryOut> getRowInfoByPurchaseNumber(Long number);
 
 }
